refactor: 프렌차이즈 엔티티 조회용 클래스 생성
